rename realtime gui
