+++ docs/cn/Changelog_CN.md
@@ -1,6 +1,6 @@
 ### 20231006更新
 
-我们制作了一个用于实时变声的界面go-realtime-gui.bat/gui_v1.py（事实上早就存在了），本次更新重点也优化了实时变声的性能。对比0813版：
+我们制作了一个用于实时变声的界面go-realtime-gui.bat/realtime-gui.py（事实上早就存在了），本次更新重点也优化了实时变声的性能。对比0813版：
  - 1、优优化界面操作：参数热更新（调整参数不需要中止再启动），懒加载模型（已加载过的模型不需要重新加载），增加响度因子参数（响度向输入音频靠近）
  - 2、优化自带降噪效果与速度
  - 3、大幅优化推理速度
