+++ go-realtime-gui-dml.bat
@@ -1,2 +1,2 @@
-runtime\python.exe gui_v1.py --pycmd runtime\python.exe --dml
+runtime\python.exe realtime-gui.py --pycmd runtime\python.exe --dml
 pause
