+++ go-realtime-gui.bat
@@ -1,2 +1,2 @@
-runtime\python.exe gui_v1.py
+runtime\python.exe realtime-gui.py
 pause
