+++ i18n/scan_i18n.py
@@ -38,7 +38,7 @@ for filename in glob.iglob("**/*.py", recursive=True):
 code_keys = set(strings)
 """
 n_i18n.py
-gui_v1.py 26
+realtime-gui.py 26
 app.py 16
 infer-web.py 147
 scan_i18n.py 0
